deploymentsample/requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
bs4==0.0.1
cachetools==4.1.0
cbor2==5.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
EasyProcess==0.3
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-Hashing==1.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gast==0.2.2
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
graphviz==0.14
grpcio==1.28.1
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
httplib2==0.18.1
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.3.1
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.0.0
playsound==1.2.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
pubnub==3.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycryptodomex==3.9.7
pydot==1.2.3
pyenchant==3.0.1
PyGetWindow==0.0.8
Pygments==2.6.1
pylint==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
PyMsgBox==1.0.7
pyparsing==2.4.6
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
pytesseract==0.3.4
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
SpeechRecognition==3.8.1
SQLAlchemy==1.3.15
svgwrite==1.3.1
sympy==1.5.1
tb-nightly==1.15.0a20190806
tensorboard==2.1.1
tensorflow==2.0.0rc1
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019080601
tornado==6.0.4
traitlets==4.3.3
Tree==0.2.4
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
zipp==3.1.0
